backend/requirements.txt
django==3.0.5
flake8==3.8.4
djangorestframework==3.11.0
drf-extra-fields==3.1.1
djoser==2.1.0
djangorestframework-simplejwt==4.6.0
django-filter==2.4.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
pillow==9.0.0